Refactor render map element by separate files
